requirements.txt
flask
requests
pillow
pypdf
reportlab
